fix(spoonacular): items observable watches active list for changes
Corrects an issue that occurred because the items observable is more persistent in a service. It was no longer being refreshed when moving from the lists page to the items page. So added a combineLatest for the collection ref and the active list subject so the observable logic will re-run when the active list changes.